+++ pkg/athena/datasource_test.go
@@ -52,68 +52,3 @@ func TestConnection_applySettings(t *testing.T) {
 		})
 	}
 }
-
-func TestConnection_getRegionKey(t *testing.T) {
-	tests := []struct {
-		description string
-		settings    *models.AthenaDataSourceSettings
-		region      string
-		catalog     string
-		expected    string
-	}{
-		{
-			"undefined region",
-			&models.AthenaDataSourceSettings{AWSDatasourceSettings: awsds.AWSDatasourceSettings{}},
-			"",
-			"",
-			"default-default",
-		},
-		{
-			"default region",
-			&models.AthenaDataSourceSettings{AWSDatasourceSettings: awsds.AWSDatasourceSettings{}},
-			"default",
-			"",
-			"default-default",
-		},
-		{
-			"same region",
-			&models.AthenaDataSourceSettings{
-				AWSDatasourceSettings: awsds.AWSDatasourceSettings{
-					DefaultRegion: "foo",
-				},
-			},
-			"foo",
-			"",
-			"default-default",
-		},
-		{
-			"different region",
-			&models.AthenaDataSourceSettings{
-				AWSDatasourceSettings: awsds.AWSDatasourceSettings{
-					Region: "foo",
-				},
-			},
-			"foo",
-			"",
-			"foo-default",
-		},
-		{
-			"different catalog",
-			&models.AthenaDataSourceSettings{
-				AWSDatasourceSettings: awsds.AWSDatasourceSettings{
-					Region: "foo",
-				},
-			},
-			"",
-			"foo",
-			"default-foo",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.description, func(t *testing.T) {
-			if res := getConnectionKey(tt.settings, tt.region, tt.catalog); res != tt.expected {
-				t.Errorf("unexpected result %v", res)
-			}
-		})
-	}
-}
